Fix windows build (static_cast missing).
TBR=ishell@chromium.org

Review URL: https://codereview.chromium.org/714413005

Cr-Commit-Position: refs/heads/master@{#25278}
git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@25278 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/string-builder.h
@@ -325,7 +325,7 @@ class IncrementalStringBuilder {
       while (*u != '\0') Append(*(u++));
     }
 
-    int written() { return cursor_ - start_; }
+    int written() { return static_cast<int>(cursor_ - start_); }
 
    private:
     DestChar* start_;
